requirements.txt
fastapi==0.104.1
jinja2==3.1.2
passlib==1.7.4
pydantic==2.5.2
python-jose==3.3.0
python-multipart==0.0.6
uvicorn==0.24.0.post1
python-qbittorrent==0.4.3
the_python_bay==1.2.3